Create preprocessing_and_qc_BASH
BASH code for second part of methods and materials for thesis. 
"""
To merge the “MEGA” and “Neuro” chips, PLINK (version 1.9, build 7.7, released October 22, 2024, 64-bit Linux version) was used. PLINK is  a bioinformatics toolset that is designed to perform tasks such as merging genotype data based on SNP call rate. 
First, a file called “keep_snps” was created, which skipped the header line for file 2, builds an associative array and for each SNP, only keeps the one with the lower missingness rate if found in both files, and then outputs the list of all unique SNPs in a tab separated text file. 
This was used to then run the following Bash code on the Linux command line: 
Which first filters both Chips to find the appropriate SNPs, and then merges them to create a new binary dataset (.bed, .bim, .fam). The .bed file 
Quality control was done in 4 separate steps: 
1.	Removing SNPs with high missingness 
Removing SNPs missing in >5% of individuals was done with the following code: 
2.	Removing SNPs with low minor allele frequency (MAF) 
Removing rare SNPs (maf < %1): 

3.	Performing Hardy-Weinberg Equilibrium (HWE) filtering 
Removing SNPs deviating from HWE: 

4.	Removing individuals with high missingness 
Removing samples with >5% missing genotypes. 
"""