Create FUMA_to_cytoscape_R
This is the code for the "Materials and Methods: Part VI. Cytoscape" section of the thesis. This code was used for the creation of the Cytoscape gene-gene interaction networks and the text file only containing the Ensembl Ids for the mapped genes (which was then subsequently used to create the STRING protein-protein interaction networks in Cytoscape).